@@ -7,37 +7,37 @@
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
     "test": "ng test --code-coverage",
-    "http-server": "http-server -p 8080 -c-1 dist/spiewnigator-angular",
+    "http-server": "npx http-server -p 8080 -c-1 dist/spiewnigator-angular/browser",
     "build-deploy": "ng build --base-href=/spiewnigator-dev/",
     "lint": "ng lint",
     "serve-coverage": "http-server -p 9000 -c-1 coverage/spiewnigator-angular"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.11",
-    "@angular/cdk": "^17.3.10",
-    "@angular/common": "^17.3.11",
-    "@angular/compiler": "^17.3.11",
-    "@angular/core": "^17.3.11",
-    "@angular/forms": "^17.3.11",
-    "@angular/material": "^17.3.10",
-    "@angular/platform-browser": "^17.3.11",
-    "@angular/platform-browser-dynamic": "^17.3.11",
-    "@angular/router": "^17.3.11",
-    "@angular/service-worker": "^17.3.11",
+    "@angular/animations": "^18.2.8",
+    "@angular/cdk": "^18.2.8",
+    "@angular/common": "^18.2.8",
+    "@angular/compiler": "^18.2.8",
+    "@angular/core": "^18.2.8",
+    "@angular/forms": "^18.2.8",
+    "@angular/material": "^18.2.8",
+    "@angular/platform-browser": "^18.2.8",
+    "@angular/platform-browser-dynamic": "^18.2.8",
+    "@angular/router": "^18.2.8",
+    "@angular/service-worker": "^18.2.8",
     "rxjs": "~6.6.0",
     "tslib": "^2.6.3",
-    "zone.js": "^0.14.7"
+    "zone.js": "^0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.8",
-    "@angular-eslint/builder": "17.5.2",
-    "@angular-eslint/eslint-plugin": "17.5.2",
-    "@angular-eslint/eslint-plugin-template": "17.5.2",
-    "@angular-eslint/schematics": "17.5.2",
-    "@angular-eslint/template-parser": "17.5.2",
-    "@angular/cli": "^17.3.8",
-    "@angular/compiler-cli": "^17.3.11",
+    "@angular-devkit/build-angular": "^18.2.8",
+    "@angular-eslint/builder": "18.3.1",
+    "@angular-eslint/eslint-plugin": "18.3.1",
+    "@angular-eslint/eslint-plugin-template": "18.3.1",
+    "@angular-eslint/schematics": "18.3.1",
+    "@angular-eslint/template-parser": "18.3.1",
+    "@angular/cli": "^18.2.8",
+    "@angular/compiler-cli": "^18.2.8",
     "@types/jasmine": "~3.8.0",
     "@types/node": "^12.20.23",
     "@typescript-eslint/eslint-plugin": "^7.2.0",
@@ -49,6 +49,6 @@
     "karma-coverage": "^2.2.1",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "^5.4.5"
+    "typescript": "~5.4.5"
   }
 }
