@@ -1,8 +1,9 @@
 import { fakeAsync, TestBed, tick } from '@angular/core/testing';
 
 import { SongProviderService } from './song-provider.service';
-import { HttpClientTestingModule , HttpTestingController } from '@angular/common/http/testing'
+import { HttpTestingController, provideHttpClientTesting } from '@angular/common/http/testing'
 import { SongRaw } from '../model/song';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('SongProviderService', () => {
   let service: SongProviderService;
@@ -15,10 +16,9 @@ describe('SongProviderService', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [
-        HttpClientTestingModule
-      ]
-    });
+    imports: [],
+    providers: [provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+});
     service = TestBed.inject(SongProviderService);
     httpMock = TestBed.inject(HttpTestingController);
   });
